requirements.txt
numpy
matplotlib
pillow
opencv
